Moves date-publication tests
